pyproject.toml
[tool.poetry]
name = "dagster-serde"
version = "0.0.0"
description = "Dagster library for ser/de of data"
authors = [
    "Daniel Gafni <danielgafni16@gmail.com>"
]
readme = "README.md"
packages = [{include = "dagster_serde"}]
repository = "https://github.com/danielgafni/dagster-serde"
keywords = [
    "dagster",
    "serde",
    "ETL",
]
classifiers = [
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
dagster = ">=1.3.5"
typing-extensions = ">=4.7.1"
pyserde = ">=0.12.3"
orjson = ">=3.9.8"


[tool.poetry.extras]


[tool.poetry.group.dev.dependencies]
hypothesis = ">=4.0.0"
pytest = "^7.3.1"
deepdiff = "^6.3.0"
isort = "^5.12.0"
ruff = "^0.0.272"
pyright = "^1.1.313"
tox = "^4.6.0"
tox-gh = "^1.0.0"
pre-commit = "^3.3.2"
dagit = "^1.3.9"
black = "^23.3.0"
pytest-cases = "^3.6.14"
blacken-docs = "^1.16.0"
pytest-rerunfailures = "^12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry-dynamic-versioning]
enable = true
strict = false
vcs = "git"
style = "pep440"
dirty = true
bump = true
metadata = false

[tool.poetry-dynamic-versioning.substitution]
files = [
    "pyproject.toml",
    "dagster_serde/_version.py"
]

[tool.pytest.ini_options]
log_cli = true
log_level = "INFO"

[tool.isort]
profile = "black"
line_length = 120
src_paths = ["dagster_serde", "tests"]

[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.pytest_cache
  | \.ruff_cache
  | \.venv
  | build
  | dist
  )/
'''

[tool.ruff]
line-length = 120
src = [
    "dagster_serde",
    "tests"
]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
[tool.ruff.isort]
known-first-party = ["dagster_serde", "tests"]

[tool.pyright]
reportPropertyTypeMismatch = true
reportImportCycles = true
reportWildcardImportFromLibrary = true
reportUntypedFunctionDecorator = true
reportUntypedClassDecorator = true

include = [
    "dagster_serde",
    "tests"
]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
